Add home button and stack for back navigation
The "stack" I created is more of a set of arrays that is populated with
data when a user navigates through the app's pages. We can then use these
arrays to display what the user previously looked at when they choose to
go back in thier navigation history. The Home button lets the user return
to the home screen from wherever they are in the app.