Start adding location functionality
I was able to get the location api working, by reloading the store when
user location is turned on. The only thing I need to finish is fixing the
map marker loading and vendcount when userlocation is turned off. The data
is sorted correctly, but those two items are not set correctly.