Display educational information in spec pages
The info window will now navigate you to the specific page with the
correct informational data that pertains to the first story. We still need
to load the correct story data for the product seleced in the detail page.
We also need real data from Mark so that we can make sure we are showing
all of the pertinet data, and so that we can format the data display.