Fix for use with device after repo split
I fixed this branch to work on the device in order to test that the map
markers were populated correctly, because after the repo split, the app on
the device was not able to access the marker image files. After I fixed
this branch so that it would work on the divice, I discovered that the
map markers were not showing up on the device at this point in development
either.